Refactor ml_dataset_new into modular structure
