+++ project/modules/machine_learning/ml_dataset_new/__init__.py
@@ -1,0 +1,26 @@
+from .core import (
+    MLResource,
+    MLResourceMetadata,
+    MLResourceStorage,
+    MLOutputCollection,
+    MLOutputCollectionStorage,
+    MLAssetsContainer,
+    MLAssetsMetadata,
+    MLAssetsContainerStorage,
+    MLDataset,
+    MLDatasetStorage,
+)
+
+__all__ = [
+    "MLResource",
+    "MLResourceMetadata",
+    "MLResourceStorage",
+    "MLOutputCollection",
+    "MLOutputCollectionStorage",
+    "MLAssetsContainer",
+    "MLAssetsMetadata",
+    "MLAssetsContainerStorage",
+    "MLDataset",
+    "MLDatasetStorage",
+]
+
